refactor: move `defaultProps` to Wrap.js
